move back to root of repo and update setup.py file to include it
